# Prune CUB's ChainedPolicy by __CUDA_ARCH_LIST__