--- a/thrust/testing/transform_input_output_iterator.cu
+++ b/thrust/testing/transform_input_output_iterator.cu
@@ -7,8 +7,16 @@
 
 #include <unittest/unittest.h>
 
+// There is an unfortunate miscompilation of the gcc-13 vectorizer leading to OOB writes
+// Adding this attribute suffices that this miscompilation does not appear anymore
+#if defined(_CCCL_COMPILER_GCC) && __GNUC__ >= 13
+#  define THRUST_DISABLE_BROKEN_GCC_VECTORIZER __attribute__((optimize("no-tree-vectorize")))
+#else // defined(_CCCL_COMPILER_GCC) && __GNUC__ >= 13
+#  define THRUST_DISABLE_BROKEN_GCC_VECTORIZER
+#endif // defined(_CCCL_COMPILER_GCC) && __GNUC__ >= 13
+
 template <class Vector>
-void TestTransformInputOutputIterator()
+THRUST_DISABLE_BROKEN_GCC_VECTORIZER void TestTransformInputOutputIterator()
 {
   using T = typename Vector::value_type;
 
@@ -52,7 +60,7 @@ void TestTransformInputOutputIterator()
 DECLARE_VECTOR_UNITTEST(TestTransformInputOutputIterator);
 
 template <class Vector>
-void TestMakeTransformInputOutputIterator()
+THRUST_DISABLE_BROKEN_GCC_VECTORIZER void TestMakeTransformInputOutputIterator()
 {
   using T = typename Vector::value_type;