Motivated by @gevtushenko and @elstehle explaining to me why CUB instantiates so many kernels and why having many tuning policies is bad, here is a mitigation: When the macro `__CUDA_ARCH_LIST__` is available, we know at compile time what runtime values the ptx version can have, so we can prune the number of dispatches CUB generates from the tuning policies to only those versions. This should give us faster compilation and allow us to use tuning policies more liberally.

Compile time and binary size of `cub.example.device.radix_sort` before and after:
```
before:
    ARCH=50;60;70;80;90: 23.826s 3886008B
    ARCH=86:              8.462s 1685520B
after:
    ARCH=50;60;70;80;90: 23.646s 3877904B
    ARCH=86:              6.095s 1232912B
```