@bernhardmgruber 2024-08-02
The unit tests now list all virtual architectures, since the list was shorter than I expected.