@bernhardmgruber 2024-08-14
I reworked the feature to now only ever instantiate to the PTX versions that appear in `__CUDA_ARCH_LIST__`, which is even better.