@davidwendt 2024-09-04
Per @elstehle  request, I tested this successfully with libcudf 24.10 without our current scan-tuning patch and this worked well us.